+++ docker/aplikacja/Dockerfile
@@ -1,0 +1,5 @@
+FROM frolvlad/alpine-oraclejdk8:slim
+ADD app.jar app.jar
+RUN sh -c 'touch /app.jar'
+ENTRYPOINT ["java","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar"]
+
