+++ docker/mantisbt/docker-compose.yml
@@ -1,0 +1,17 @@
+mantisbt:
+  image: vimagick/mantisbt:latest
+  ports:
+    - "8989:80"
+  links:
+    - mysql
+  restart: always
+
+mysql:
+  image: mysql:latest
+  environment:
+    - MYSQL_ROOT_PASSWORD=root
+    - MYSQL_DATABASE=bugtracker
+    - MYSQL_USER=mantisbt
+    - MYSQL_PASSWORD=mantisbt
+  restart: always
+
