+++ src/main/java/pik/Application.java
@@ -1,15 +1,16 @@
-package hello;
+package pik;
 
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
 import org.springframework.boot.context.embedded.ErrorPage;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.HttpStatus;
 
 @SpringBootApplication
+@EnableAutoConfiguration
 public class Application {
 	@Bean
 	public EmbeddedServletContainerCustomizer containerCustomizer() {
