+++ src/main/java/pik/dto/SequenceId.java
@@ -1,4 +1,4 @@
-package dto;
+package pik.dto;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
