+++ src/main/java/pik/exceptions/SequenceException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package pik.exceptions;
 
 public class SequenceException extends RuntimeException {
