+++ src/main/java/pik/repositories/UserRepository.java
@@ -1,6 +1,6 @@
-package repositories;
+package pik.repositories;
 
-import dto.UserInfo;
+import pik.dto.UserInfo;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.List;
 
@@ -8,6 +8,5 @@ public interface UserRepository extends MongoRepository <UserInfo,String> {
     List<UserInfo> findByFirstName(String firstName);
     List<UserInfo> findByLastName (String lastName);
     UserInfo findByEmail(String eMail);
-    UserInfo findByUserId(String id);
-    UserInfo findByUserName(String username);
+    UserInfo findByuserId(String id);
 }
